Remove password check endpoint

+++ frontend/src/reducers/usersReducer.js
@@ -45,18 +45,13 @@ export const getUsers = (token) => {
 export const deleteUser = (user, token, confirmText, handleClose, history) => {
     return async (dispatch, getState) => {
         const library = getState()?.language?.library.frontend.users.reducer
-        if (confirmText && handleClose) {
-            const pass = await userService.checkPassword({ id: user.id, token: token, confirmText: confirmText })
-            if (pass.status !== 200) {
-                dispatch(setNotification({ message: pass.error, success: false, show: true }))
-                return
-            }
-        }
-        const res = await userService.deleteUser(user.id, token)
+        const res = await userService.deleteUser(user.id, confirmText, token)
         if (res.status !== 204) {
             dispatch(setNotification({ message: res.error, success: false, show: true }))
         } else {
-            dispatch(getCredits(token))
+            if(!confirmText) {
+                dispatch(getCredits(token))
+            }
             if (handleClose) {
                 handleClose()
                 dispatch(logout(history))
